Update and rename app.py to bot.py
- Renamed app.py to bot.py for better clarity.
- Updated the CSS with improvements in functionality.
- Ensured compatibility with the existing framework.
